split raw-linearizability
